+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fix
 
 * Change order of event filtering mechanisms (#2001)
+* Only send session update for dropped events if state changed (#2002)
 
 ## 6.0.0-beta.2